Update and rename BounderR Module 1-1 Import Coordinates_v20160816.R to BounderR Module 1-1 Import Coordinates_v20170201.R
Altered for compatibility with newer versions of CAD-Bounder (e.g. v20170201). Definition of 'bb' variable for back-projected frames was not accurate anymore in newer versions (because the structure of the coordinates table had changed between versions). Changed for 'mm' within this script for internal handling (for backwards compatibility with versions which contain a mismatched number of rows between Filopodia and Coordinates tables. 

'bb' is now explicitly defined on top of Module 1  (safer and more transparent!)